DCAC-78: Repackage, rationalise code following merge.
