+++ src/main/java/uk/gov/companieshouse/itemgroupworkflowapi/ItemGroupWorkflowApiApplication.java
@@ -3,13 +3,15 @@ package uk.gov.companieshouse.itemgroupworkflowapi;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import static uk.gov.companieshouse.itemgroupworkflowapi.environment.EnvironmentVariablesChecker.allRequiredEnvironmentVariablesPresent;
+
 @SpringBootApplication
 public class ItemGroupWorkflowApiApplication {
 
-    public static final String NAMESPACE = "item-group-workflow-api";
-
     public static void main(String[] args) {
-        SpringApplication.run(ItemGroupWorkflowApiApplication.class, args);
+        if(allRequiredEnvironmentVariablesPresent()) {
+            SpringApplication.run(ItemGroupWorkflowApiApplication.class, args);
+        }
     }
 
 }
