+++ src/main/java/uk/gov/companieshouse/itemgroupworkflowapi/controller/ItemGroupController.java
@@ -4,16 +4,26 @@ import jakarta.json.JsonMergePatch;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.companieshouse.api.error.ApiError;
+import uk.gov.companieshouse.itemgroupworkflowapi.logging.LoggingUtils;
+import uk.gov.companieshouse.itemgroupworkflowapi.model.ItemGroupCreate;
+import uk.gov.companieshouse.itemgroupworkflowapi.model.ItemGroupJsonPayload;
+import uk.gov.companieshouse.itemgroupworkflowapi.service.ItemGroupsService;
+import uk.gov.companieshouse.itemgroupworkflowapi.validator.CreateItemValidator;
 import uk.gov.companieshouse.itemgroupworkflowapi.validator.PatchItemRequestValidator;
-import uk.gov.companieshouse.logging.Logger;
 
 import java.util.List;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import static org.springframework.http.HttpStatus.CONFLICT;
+import static org.springframework.http.HttpStatus.CREATED;
+import static uk.gov.companieshouse.itemgroupworkflowapi.logging.LoggingUtils.CREATE_ITEM_GROUP_ERROR_PREFIX;
+import static uk.gov.companieshouse.itemgroupworkflowapi.logging.LoggingUtils.CREATE_ITEM_GROUP_REQUEST;
+import static uk.gov.companieshouse.itemgroupworkflowapi.logging.LoggingUtils.CREATE_ITEM_GROUP_RESPONSE;
 import static uk.gov.companieshouse.itemgroupworkflowapi.util.Constants.REQUEST_ID_HEADER_NAME;
 import static uk.gov.companieshouse.itemgroupworkflowapi.util.PatchMediaType.APPLICATION_MERGE_PATCH_VALUE;
 
@@ -23,13 +33,41 @@ public class ItemGroupController {
     private static final String PATCH_ITEM_URI =
             "${uk.gov.companieshouse.itemgroupworkflowapi.patchitem}";
 
-    private final Logger logger;
+    private final LoggingUtils logger;
+    private final ItemGroupsService itemGroupsService;
+    private final CreateItemValidator createItemValidator;
 
     private final PatchItemRequestValidator patchItemRequestValidator;
 
-    public ItemGroupController(Logger logger, PatchItemRequestValidator patchItemRequestValidator) {
+    public ItemGroupController(LoggingUtils logger,
+                               ItemGroupsService itemGroupsService,
+                               CreateItemValidator createItemValidator,
+                               PatchItemRequestValidator patchItemRequestValidator) {
         this.logger = logger;
+        this.itemGroupsService = itemGroupsService;
+        this.createItemValidator = createItemValidator;
         this.patchItemRequestValidator = patchItemRequestValidator;
+    }
+
+    @PostMapping("${uk.gov.companieshouse.itemgroupworkflowapi.createitemgroup}")
+    public ResponseEntity<Object> createItemGroup(final @RequestBody ItemGroupJsonPayload itemGroupJsonPayload) {
+        List<String> errors = createItemValidator.validateCreateItemPayload(itemGroupJsonPayload);
+
+        if (!errors.isEmpty()) {
+            return buildValidationResponse(BAD_REQUEST.value(), errors, itemGroupJsonPayload);
+        }
+
+        try {
+            if (itemGroupsService.doesItemGroupExist(itemGroupJsonPayload)) {
+                return buildItemAlreadyExistsResponse(itemGroupJsonPayload);
+            }
+
+            final ItemGroupCreate savedItem = itemGroupsService.createItemGroup(itemGroupJsonPayload);
+            return buildCreateSuccessResponse(savedItem);
+
+        } catch (Exception ex) {
+            return buildErrorResponse(BAD_REQUEST.value(), ex, itemGroupJsonPayload);
+        }
     }
 
     @PatchMapping(path = PATCH_ITEM_URI, consumes = APPLICATION_MERGE_PATCH_VALUE)
@@ -40,14 +78,14 @@ public class ItemGroupController {
             final @RequestHeader(REQUEST_ID_HEADER_NAME) String requestId) {
 
         // TODO DCAC-78 Use structured logging
-        logger.info("patchItem(" + mergePatchDocument +
+        logger.getLogger().info("patchItem(" + mergePatchDocument +
                 ", " + itemGroupId + ", " + itemId + ", " + requestId + ") called.");
 
         final List<ApiError> errors = patchItemRequestValidator.getValidationErrors(mergePatchDocument);
         if (!errors.isEmpty()) {
             // TODO DCAC-78 Use structured logging
             //  logErrorsWithStatus(logMap, errors, BAD_REQUEST);
-            logger.error("update item request had validation errors " + errors);
+            logger.getLogger().error("update item request had validation errors " + errors);
             return ApiErrors.errorResponse(BAD_REQUEST, errors);
         }
 
@@ -62,6 +100,37 @@ public class ItemGroupController {
         // TODO DCAC-78 Build response DTO and return it as body
 
         return ResponseEntity.ok().build();
+    }
+
+
+    private ResponseEntity<Object> buildCreateSuccessResponse(final ItemGroupCreate savedItem) {
+        return ResponseEntity.status(CREATED).body(savedItem);
+    }
+
+    private ResponseEntity<Object> buildValidationResponse(final int statusCode,
+                                                           final List<String> errors,
+                                                           final ItemGroupJsonPayload dto){
+        final var map = logger.createLogMap();
+        map.put(CREATE_ITEM_GROUP_REQUEST, dto);
+        final ResponseEntity<Object> response = ResponseEntity.status(statusCode).body(errors);
+        map.put(CREATE_ITEM_GROUP_RESPONSE, response);
+        logger.getLogger().error(CREATE_ITEM_GROUP_ERROR_PREFIX + errors, map);
+        return response;
+    }
+
+    private ResponseEntity<Object> buildItemAlreadyExistsResponse(final ItemGroupJsonPayload dto) {
+        return ResponseEntity.status(CONFLICT).body(dto);
+    }
+
+    private ResponseEntity<Object> buildErrorResponse(final int statusCode,
+                                                      final Exception ex,
+                                                      final ItemGroupJsonPayload dto){
+        final var map = logger.createLogMap();
+        map.put(CREATE_ITEM_GROUP_REQUEST, dto);
+        final ResponseEntity<Object> response = ResponseEntity.status(statusCode).body(ex.getMessage());
+        map.put(CREATE_ITEM_GROUP_RESPONSE, response);
+        logger.getLogger().error(CREATE_ITEM_GROUP_ERROR_PREFIX + ex.getMessage(), map);
+        return response;
     }
 
 }
