+++ src/main/java/uk/gov/companieshouse/itemgroupworkflowapi/repository/ItemGroupsRepository.java
@@ -1,8 +1,8 @@
-package uk.gov.companieshouse.item.group.workflow.api.repository;
+package uk.gov.companieshouse.itemgroupworkflowapi.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import uk.gov.companieshouse.item.group.workflow.api.model.ItemGroupCreate;
+import uk.gov.companieshouse.itemgroupworkflowapi.model.ItemGroupCreate;
 
 @Repository
 public interface ItemGroupsRepository extends MongoRepository<ItemGroupCreate, String> {
