--- src/test/java/uk/gov/companieshouse/item/group/workflow/api/ItemGroupCreateWorkflowApiApplicationTests.java
@@ -1,14 +1,0 @@
-package uk.gov.companieshouse.item.group.workflow.api;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ItemGroupCreateWorkflowApiApplicationTests {
-
-    @SuppressWarnings("squid:S2699") // at least one assertion
-    @Test
-    void contextLoads() {
-    }
-
-}
